backend/requirements.txt
# Flask Core
Flask==2.3.3
Flask-Session==0.5.0
Flask-Compress==1.13
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5

# AI Integration
openai==1.3.0
google-generativeai==0.3.2

# RAG & Search
chromadb==0.4.15
sentence-transformers==2.2.2
rank-bm25==0.2.2

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Google APIs
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0

# Utilities
pathlib2==2.3.7
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-flask==1.2.0
black==23.11.0
flake8==6.1.0

# Production
gunicorn==21.2.0
waitress==2.1.2
